chore: remove deprecated configuration and script files, add new setup-in-workspace script for argent installation
