--- .vscode/launch.json
@@ -1,72 +1,0 @@
-{
-  "version": "0.2.0",
-  "inputs": [
-    {
-      "id": "projectPath",
-      "type": "promptString",
-      "description": "Path to the project where Argent should be installed"
-    }
-  ],
-  "configurations": [
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Tools Server",
-      "runtimeArgs": ["-r", "ts-node/register"],
-      "program": "${workspaceFolder}/packages/tool-server/src/index.ts",
-      "cwd": "${workspaceFolder}/packages/tool-server",
-      "console": "integratedTerminal",
-      "env": {
-        "PORT": "3001"
-      }
-    },
-    {
-      "type": "chrome",
-      "request": "launch",
-      "name": "UI (Chrome)",
-      "url": "http://localhost:5173",
-      "webRoot": "${workspaceFolder}/packages/ui",
-      "preLaunchTask": "start-vite"
-    },
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Tools Server (built)",
-      "program": "${workspaceFolder}/packages/tool-server/dist/index.js",
-      "cwd": "${workspaceFolder}/packages/tool-server",
-      "console": "integratedTerminal",
-      "env": {
-        "PORT": "3001"
-      },
-      "preLaunchTask": "build-tools"
-    },
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Install Argent in Project",
-      "program": "${workspaceFolder}/scripts/setup-project.cjs",
-      "args": ["${input:projectPath}"],
-      "console": "integratedTerminal",
-      "internalConsoleOptions": "neverOpen"
-    },
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Argent Agent Debug",
-      "runtimeArgs": ["-r", "ts-node/register"],
-      "program": "${workspaceFolder}/packages/tool-server/src/index.ts",
-      "cwd": "${workspaceFolder}/packages/tool-server",
-      "console": "integratedTerminal",
-      "env": {
-        "PORT": "3001"
-      },
-      "preLaunchTask": "setup-dev-in-project"
-    }
-  ],
-  "compounds": [
-    {
-      "name": "Full (Tools + UI)",
-      "configurations": ["Tools Server", "UI (Chrome)"]
-    }
-  ]
-}
