--- .vscode/tasks.json
@@ -1,35 +1,0 @@
-{
-  "version": "2.0.0",
-  "inputs": [
-    {
-      "id": "projectPath",
-      "type": "promptString",
-      "description": "Absolute path to the target project folder"
-    }
-  ],
-  "tasks": [
-    {
-      "label": "build-tools",
-      "type": "shell",
-      "command": "npm run build",
-      "options": {
-        "cwd": "${workspaceFolder}"
-      },
-      "problemMatcher": ["$tsc"],
-      "group": "build"
-    },
-    {
-      "label": "setup-dev-in-project",
-      "type": "shell",
-      "command": "node \"${workspaceFolder}/scripts/setup-dev-in-project.cjs\" \"${input:projectPath}\"",
-      "options": {
-        "cwd": "${workspaceFolder}"
-      },
-      "problemMatcher": [],
-      "presentation": {
-        "reveal": "always",
-        "panel": "shared"
-      }
-    }
-  ]
-}
